refactor: introduce typed Config, S3 transfermanager uploads, and embedded templates
- Replace package-level config vars/init() with config.LoadConfig() returning *Config
- Move NewfsClient/NewS3Client onto App so storage clients read app.Config
- Switch S3 uploads to transfermanager.UploadObject; verify bucket via HeadBucket on connect
- Enforce max upload size via limitUploadSize middleware
- Embed HTML templates with go:embed instead of loading from disk
- Normalize handler error responses to consistent status codes
- Regenerate sqlc queries (int64 -> int) and bump Go/deps
- Add testify as a dependency in preparation for unit tests (none yet)
- Remove legacy db/sql, sqlc.yml, and unused handlers

+++ internal/database/querier.go
@@ -17,8 +17,8 @@ type Querier interface {
 	GetArchiveById(ctx context.Context, id int) (string, error)
 	GetArchiveByName(ctx context.Context, filename string) (Archive, error)
 	GetArchivePage(ctx context.Context, arg GetArchivePageParams) ([]GetArchivePageRow, error)
-	GetCountArchive(ctx context.Context) (int64, error)
-	GetCountSearch(ctx context.Context, dollar_1 sql.NullString) (int64, error)
+	GetCountArchive(ctx context.Context) (int, error)
+	GetCountSearch(ctx context.Context, dollar_1 sql.NullString) (int, error)
 	InsertFile(ctx context.Context, arg InsertFileParams) (int, error)
 	SearchArchive(ctx context.Context, arg SearchArchiveParams) ([]SearchArchiveRow, error)
 	SetEditFile(ctx context.Context, arg SetEditFileParams) error
